refactor: reformat codebase and migrate to timelines_plus package

+++ lib/view/intro/components/description_text.dart
@@ -17,7 +17,7 @@ class AnimatedDescriptionText extends StatelessWidget {
           'I\'m capable of creating excellent mobile apps, handling${Responsive.isLargeMobile(context) ? '\n' : ''}every step from ${!Responsive.isLargeMobile(context) ? '\n' : ''}concept to deployment.',
           maxLines: 2,
           overflow: TextOverflow.ellipsis,
-          style: TextStyle(color: Colors.grey, wordSpacing: 2, fontSize: value as double),
+          style: TextStyle(color: Colors.grey, wordSpacing: 2, fontSize: value),
         );
       },
     );
