+++ lib/view/intro/components/side_menu_button.dart
@@ -17,7 +17,7 @@ class MenuButton extends StatelessWidget {
               onTap: onTap,
               borderRadius: BorderRadius.circular(20),
               child: Container(
-                height: defaultPadding * 2.0 * (value as double),
+                height: defaultPadding * 2.0 * (value),
                 width: defaultPadding * 2.0 * value,
                 decoration: BoxDecoration(
                     borderRadius: BorderRadius.circular(10),
@@ -30,20 +30,19 @@ class MenuButton extends StatelessWidget {
                           color: Colors.blue.withOpacity(.5),
                           offset: const Offset(-1, -1)),
                     ]),
-                child:  Center(
-                  child: ShaderMask(
-                    shaderCallback: (bounds) {
-                      return LinearGradient(
-                              colors: [Colors.pink, Colors.blue.shade900])
-                          .createShader(bounds);
-                    },
-                    child: Icon(
-                      Icons.menu,
-                      color: Colors.white,
-                      size: defaultPadding * 1.2 * value,
-                    ),
-                  )
-                ),
+                child: Center(
+                    child: ShaderMask(
+                  shaderCallback: (bounds) {
+                    return LinearGradient(
+                            colors: [Colors.pink, Colors.blue.shade900])
+                        .createShader(bounds);
+                  },
+                  child: Icon(
+                    Icons.menu,
+                    color: Colors.white,
+                    size: defaultPadding * 1.2 * value,
+                  ),
+                )),
               ),
             );
           },
