+++ lib/view/intro/components/subtitle_text.dart
@@ -5,8 +5,13 @@ class AnimatedSubtitleText extends StatelessWidget {
   final double end;
   final String text;
   final bool gradient;
-  const AnimatedSubtitleText(
-      {super.key, required this.start, required this.end, required this.text, this.gradient=false,});
+  const AnimatedSubtitleText({
+    super.key,
+    required this.start,
+    required this.end,
+    required this.text,
+    this.gradient = false,
+  });
   @override
   Widget build(BuildContext context) {
     return TweenAnimationBuilder(
@@ -18,12 +23,20 @@ class AnimatedSubtitleText extends StatelessWidget {
           style: Theme.of(context).textTheme.headlineLarge!.copyWith(
               color: Colors.white,
               fontWeight: FontWeight.w900,
-              shadows: gradient? [
-                const Shadow(color: Colors.pink,offset: Offset(0, 2),blurRadius: 10),
-                const Shadow(color: Colors.pink,offset: Offset(0, -2),blurRadius: 10),
-              ] :[] ,
+              shadows: gradient
+                  ? [
+                      const Shadow(
+                          color: Colors.pink,
+                          offset: Offset(0, 2),
+                          blurRadius: 10),
+                      const Shadow(
+                          color: Colors.pink,
+                          offset: Offset(0, -2),
+                          blurRadius: 10),
+                    ]
+                  : [],
               height: 0,
-              fontSize: value as double),
+              fontSize: value),
         );
       },
     );
